EcommerceApi/requirements.txt.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
httpie==2.4.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
pipenv==2020.11.15
psycopg2==2.8.6
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PySocks==1.7.1
pytz==2020.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
wrapt==1.12.1